Basic skills HUD is now functional. A player can cycle between and activate skills now.

+++ wlr/OGUR/OGUR/Sprites/AnimatedTexture.cs
@@ -86,10 +86,10 @@ namespace OGUR.Sprites
             return m_position;
         }
 
-        public void SetPosition(int x, int y)
+        public void SetPosition(double x, double y)
         {
-            m_position.X = x;
-            m_position.Y = y;
+            m_position.X = (int)x;
+            m_position.Y = (int)y;
         }
     }
 }
